email generation as well as editSubscription verification

+++ db/twilio.js
@@ -1,59 +1,79 @@
-const twilio = require('twilio');
-const db = require('../db/index');    
-const sgMail     = require('@sendgrid/mail');
+const twilio            = require('twilio');
+const db                = require('../db/index');    
+const { emailTemplate } = require('../db/generateEmailTemplate');
+const sgMail            = require('@sendgrid/mail');
 sgMail.setApiKey(process.env.SENDGRID_API_KEY);
 // twilio CONFIG *********
-const client = new twilio(process.env.accountSid, process.env.authToken);
+const client            = new twilio(process.env.accountSid, process.env.authToken);
 
 module.exports = {
     // daily updates for users 
     async dailyUpdate () {
         try {
-            // come up with a better QUERY HERE 
-            let screeningSql = 'SELECT name, title, link, showtime FROM theatre INNER JOIN screening ON theatre.id = screening.theatre_id ORDER BY name;';
-            const { rows } = await db.query(screeningSql);
+            let screeningSql = 'SELECT * FROM theatre INNER JOIN screening ON theatre.id = screening.theatre_id ORDER BY name;';
+            const result = await db.query(screeningSql);
+            const screenings = result.rows;
 
-            let userSql = 'SELECT * FROM "user" WHERE user.paused <= $1';
+            let userSql = 'SELECT * FROM "user" WHERE paused <= $1';
             let userParams = [ Date.now() ];
             const results = await db.query(userSql, userParams);
             const users = results.rows;
             
+            // generate email content
+            const content = emailTemplate(screenings);
+
             for(const user of users) {
-                if(user.verify_token_expires < user.created_date) {
+                if(user.verify_token_expires && user.verify_token_expires < user.created_date) {
                     let sql = 'DELETE FROM "user" WHERE id = $1';
                     let params = [user.id];
 
                     await db.query(sql, params);
-                }
-                if(user.textupdate) {
+                }    
+                if(user.text_update) {
                     // SMS 
-                    for(const row of rows) {
+                    for(const screening of screenings) {
                         const message = await client.messages.create({
-                            body: `${row.name}, ${row.title}, ${row.link}, ${row.showtime}
-                            
-                            Unsubscribe or snooze here: http://${req.headers.host}/users/pause`,
                             to: user.number,  
-                            from: process.env.twilioNumber 
+                            from: process.env.twilioNumber,
+                            body: `${screening.name}, ${screening.title}, ${screening.link}, ${screening.showtime}`
                         })
-                        console.log(message.sid);            
-                    }    
+                        console.log(message.sid);                
+                    }
                 } else {
-                    // EMAIL  
-                    for(const row of rows) {
-                        const msg = {
-                            to: user.email,
-                            from: `PTST Admin <${process.env.myEmail}>`,
-                            subject: 'Todays Independent Screenings Toronto!',
-                            text: `${row.name}, ${row.title}, ${row.link}, ${row.showtime}
-                            
-                            Unsubscribe or snooze here: http://${req.headers.host}/users/pause`
-                        }
-                        await sgMail.send(msg);            
-                    }    
+                    // EMAIL    
+                    const msg = {
+                        to: user.email,
+                        from: `PTST Admin <${process.env.myEmail}>`,
+                        subject: 'Todays Independent Screenings Toronto!',
+                        html: content.toString(), 
+                        text: 'hii'
+                    }
+                    await sgMail.send(msg);            
                 }
             }
         } catch (err) {
-            console.log('ERROR:', err); 
+            console.log('TWILIO ERROR:', err); 
         }
     }
 }
+// ******** mock screenings for email generation ********
+// const screenings = [
+//   {
+//       name: 'tiff',
+//       title: 'portrait of a lady on fire',
+//       link: 'www.lick_nuts.ca',
+//       showtime: [
+//           "8:00pm",
+//           "9:00am"
+//       ]
+//   },
+//   {
+//       name: 'revue',
+//       title: 'princess mananokai',
+//       link: 'www.lick_testis.ca',
+//       showtime: [
+//           "8:00am",
+//           "11:00am"
+//       ]
+//   }
+// ]
